[feat]! Update to support newer k6
Metrics were being refactored around the era of k6 0.40.0, and with 0.41
a breaking change was made that requires code updates to use.

This change updates the code in order to work with 0.41 and newer
versions of k6, at the cost of breaking compatability with 0.40 and
older versions of k6.

Since the current version of k6 is 0.53.0 at the time of writing, we
don't feel like this is too onerous.

Note:
- Open Telemetry is pinned to 0.24 instead of 0.28 / 0.30 used in k6
  0.53 due to build issues. See
  https://github.com/grafana/k6/issues/3897